Add New Object mbtaLines
Create a new Object mbtaLines with key values corresponding to the names of the
three lines required for the function stopsBetweenStations, and values as arrays
containing the names of the stops on each line. This consolidates all of the
basic information required for calculating the distance between a start station
and an end station.