Did bonus question
This was probably too many changes for one commit. I regret not committing
more often. I basically used the original assignment function, and called it
over a list of possible crossing points. The final distance was the minimum
of the different distances possible. I assumed that people want to travel the
minimum distance, and I didn't add a dictionary of actual distances between
the stops to weight them or anything like that. That could be an interesting
step though.